Restructure core pipeline
