+++ src/Juvix/Core.hs
@@ -1,10 +1,12 @@
 module Juvix.Core
   ( module Juvix.Core.Erasure,
     module Juvix.Core.Translate,
+    module Juvix.Core.Pipeline,
     module Juvix.Core.Usage,
   )
 where
 
 import Juvix.Core.Erasure
+import Juvix.Core.Pipeline
 import Juvix.Core.Translate
 import Juvix.Core.Usage
