+++ src/Juvix/Core/Erased.hs
@@ -1,10 +1,8 @@
 module Juvix.Core.Erased
   ( module Juvix.Core.Erased.Types,
     module Juvix.Core.Erased.Evaluator,
-    module Juvix.Core.Erased.Transform,
   )
 where
 
 import Juvix.Core.Erased.Evaluator
-import Juvix.Core.Erased.Transform
 import Juvix.Core.Erased.Types
