+++ src/Juvix/Core/Types.hs
@@ -1,5 +1,7 @@
 module Juvix.Core.Types where
 
+import qualified Juvix.Core.EAC.Types as EAC
+import qualified Juvix.Core.Erasure.Types as ET
 import Juvix.Library
 import Text.ParserCombinators.Parsec
 import qualified Text.ParserCombinators.Parsec.Token as Token
@@ -14,3 +16,14 @@ data Parameterisation primTy primVal
         reservedNames ∷ [String],
         reservedOpNames ∷ [String]
       }
+
+data PipelineError primTy primVal
+  = InternalInconsistencyError Text
+  | TypecheckerError Text
+  | EACError (EAC.Errors primTy primVal)
+  | ErasureError ET.ErasureError
+  deriving (Show, Generic)
+
+data PipelineLog primTy primVal
+  = LogTodo
+  deriving (Show, Generic)
