+++ test/Bohm.hs
@@ -10,6 +10,7 @@ import Juvix.Backends.Env
 import Juvix.Backends.Graph
 import Juvix.Backends.Interface
 import Juvix.Backends.Maps
+import Juvix.Bohm
 import Juvix.Bohm.Default
 import Juvix.Bohm.Parser
 import Juvix.Bohm.Translation
@@ -21,6 +22,9 @@ import Juvix.Utility
 import Juvix.Visualize.Dot
 import Juvix.Visualize.Graph
 import Text.Parsec
+
+ealToBohm ∷ ∀ primVal. RPTO primVal → Bohm
+ealToBohm = erasedCoreToBohm . erase
 
 astToNetDefault ∷ Network net ⇒ Bohm → net Lang
 astToNetDefault net = astToNet net defaultEnv
